+++ Libplanet.Net/InvalidMessageTimestampException.cs
@@ -14,25 +14,13 @@ namespace Libplanet.Net
     {
         internal InvalidMessageTimestampException(
             string message,
-            DateTimeOffset createdOffset,
-            TimeSpan? buffer,
-            DateTimeOffset currentOffset,
-            Exception innerException
-        )
-            : base(message, innerException)
-        {
-            CreatedOffset = createdOffset;
-            Buffer = buffer;
-            CurrentOffset = currentOffset;
-        }
-
-        internal InvalidMessageTimestampException(
-            string message,
+            Peer peer,
             DateTimeOffset createdOffset,
             TimeSpan? buffer,
             DateTimeOffset currentOffset)
             : base(message)
         {
+            Peer = peer;
             CreatedOffset = createdOffset;
             Buffer = buffer;
             CurrentOffset = currentOffset;
@@ -40,14 +28,16 @@ namespace Libplanet.Net
 
         protected InvalidMessageTimestampException(
             SerializationInfo info,
-            StreamingContext context
-        )
+            StreamingContext context)
             : base(info, context)
         {
+            Peer = info.GetValue<Peer>(nameof(Peer));
             CreatedOffset = info.GetValue<DateTimeOffset>(nameof(CreatedOffset));
             Buffer = info.GetValue<TimeSpan?>(nameof(Buffer));
             CurrentOffset = info.GetValue<DateTimeOffset>(nameof(CurrentOffset));
         }
+
+        internal Peer Peer { get; private set; }
 
         internal DateTimeOffset CreatedOffset { get; private set; }
 
@@ -59,6 +49,7 @@ namespace Libplanet.Net
             SerializationInfo info, StreamingContext context)
         {
             base.GetObjectData(info, context);
+            info.AddValue(nameof(Peer), Peer);
             info.AddValue(nameof(CreatedOffset), CreatedOffset);
             info.AddValue(nameof(Buffer), Buffer);
             info.AddValue(nameof(CurrentOffset), CurrentOffset);