+++ Libplanet.Net/Messages/NetMQMessageCodec.cs
@@ -113,7 +113,7 @@ namespace Libplanet.Net.Messages
                           $"the current datetime offset is " +
                           $"{currentTime.ToString(TimestampFormat, CultureInfo.InvariantCulture)}.";
                 throw new InvalidMessageTimestampException(
-                    msg, timestamp, _messageTimestampBuffer, currentTime);
+                    msg, remotePeer, timestamp, _messageTimestampBuffer, currentTime);
             }
 
             byte[] signature = remains[(int)Message.MessageFrame.Sign].ToByteArray();
@@ -136,12 +136,16 @@ namespace Libplanet.Net.Messages
                 remains[(int)Message.MessageFrame.Timestamp],
             };
 
-            var messageForVerify = headerWithoutSign.Concat(body).ToArray();
+            var messageForVerify = headerWithoutSign.Concat(body).ToByteArray();
 
-            if (!remotePeer.PublicKey.Verify(messageForVerify.ToByteArray(), signature))
+            if (!remotePeer.PublicKey.Verify(messageForVerify, signature))
             {
                 throw new InvalidMessageSignatureException(
-                    "The message signature is invalid", message);
+                    "The message signature is invalid",
+                    remotePeer,
+                    remotePeer.PublicKey,
+                    messageForVerify,
+                    signature);
             }
 
             if (!reply)